+++ src/Observers/ProductInventoryObserver.php
@@ -49,6 +49,11 @@ class ProductInventoryObserver extends AbstractProductImportObserver
             return;
         }
 
+        // query whether or not, a website ID has been specified
+        if (!$this->hasValue(ColumnKeys::WEBSITE_ID)) {
+            return;
+        }
+
         // prepare, initialize and persist the stock status/item
         $this->persistStockStatus($this->initializeStockStatus($this->prepareStockStatusAttributes()));
         $this->persistStockItem($this->initializeStockItem($this->prepareStockItemAttributes()));