+++ tests/Observers/UrlRewriteUpdateObserverTest.php
@@ -162,10 +162,13 @@ class UrlRewriteUpdateObserverTest extends \PHPUnit_Framework_TestCase
                     ->method('getHeader')
                     ->withConsecutive(
                         array(ColumnKeys::SKU),
+                        array(ColumnKeys::SKU),
                         array(ColumnKeys::URL_KEY),
+                        array(ColumnKeys::URL_KEY),
+                        array(ColumnKeys::STORE_VIEW_CODE),
                         array(ColumnKeys::STORE_VIEW_CODE)
                     )
-                    ->willReturnOnConsecutiveCalls(0, 1, 2);
+                    ->willReturnOnConsecutiveCalls(0, 0, 1, 1, 2, 2);
         $mockSubject->expects($this->any())
                     ->method('getLastEntityId')
                     ->willReturn($entityId);